Add a 'muted_channels_activity' config option
This allows you to control which activity from muted channels you will
be notified for.

The default is set to personal_highlights, as I find that the most
useful. The official client doesn't send a notification for any
activity, but it shows highlights in the sidebar, so all_highlights is
probably the closest to that. However, since there isn't any difference
between highlights from muted and non-muted channels in wee-slack, I
think getting channel highlights from muted channels is a bit much.