Fix missing nick for own messages
This was introduced in commit 51c2e76 because I removed user from the
request believing it to be unnecessary.

I think the reason I didn't experience this problem after removing it
(I'm pretty sure I tested it), is that we might receive both a message
event and a request response for the message, and we use the first of
those that arrives. The message event will always contain the user,
while the request response will not (but it's passed along if provided
as in this commit).

Fixes #612