Cleanup process_reply
There is no point in creating a SlackMessage when we're just going to
use message_json. I think this is just left over from when more of the
message was actually used.

It's not necessary to reset ws_counter. Python supports arbitrarily
large ints, and Slacks docs doesn't specify a limit (it does says that
the int should be unique for that connection, so this is more in line
with the docs). I tested the API with 2^65 and that worked fine. It
returns a string instead of an int if it's larger than 2^31 - 1 though,
so we convert it to an int to be sure.