Use highlight notifications for MPDMs
DMs with multiple persons should highlight just like DMs with one
person, instead of notifying without highlight like a channel.

This seems to have been the intention all along, but it didn't work
because the channel type was checked for mpdm, not mpim which is what
the type is set to (even though the class is called SlackMPDMChannel.